recipes-apps : add protection-relays-hmi demo
This demo is developed as part of TI Design on Protection Relaysi HMI.

A protection relay is a smart device that receives inputs, compares
them to set points, and provides outputs. Inputs can be current,
voltage, resistance, or temperature. Outputs can include visual
feedback in the form of indicator lights and/or an alphanumeric
display, communications, control warnings, alarms,  and turning
power off and on.

Signed-off-by: Manisha Agrawal <manisha.agrawal@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>