tisdk-uenv: k2g-hs-evm: fix issue with mixing bootmodes
The get_*_net helper commands use the built-in dhcp command to fetch. If we
wish to mount the rootfs over NFS, then the init_nfs helper will set
"autoload=no". This causes the "dhcp" built-in to only obtain the IP address
and not load the file, and thus the get_*_net helpers do not work as expected.

Work around this by using "tftp" instead of "dhcp", and set autoload=no and
manually trigger dhcp.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>