tisdk-uenv: k3: workaround U-Boot env saved to eMMC
The environment is saved to eMMC, but we wish that a newly created SD card
will boot without any manual configuration. To do this we will reset the
environment upon the first boot, and then create a file in the boot partition
of the SD card to mark when the first-time setup has been done.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>