sourceipk: fix package creation
Recent updates to how the rootfs is constructed has introduced image
private package feeds. The assembly of these private feeds has been
implemented in a way that it may only packages create by
"do_package_write_*" tasks.

The sourceipk class has its own implementation for creaing the
packages, so now they are not available for any image.

This change modifies the behavior of the sourceipk class so that it
can use the do_package_write tasks. This is done by staging the
sources prior to being built. As the packages may only be created
from files relative to ${D}, another tasks installs these files.
These files are then blacklisted from the sysroot so that they are
not seen by any other recipe.

Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>