conf: images: Fix errors with wificfg in WIC images
The WIC image functions use IMAGE_BOOT_FILES variable to package
all the files for the boot partition.

The wifi-oob adds a file wificfg that is required to be included
in the default and docker WIC images, but not required in the
base and tiny WIC images.

Hence remove this from conf file and add in individual recipes.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>