recipes-*: Use generic SOC family am64xx for overriding
Now that we have two machines am64xx-evm and am64xx-hs-evm,
use the generic family name "am64xx" for overriding variables that
are applicable in both cases.

Only exceptions not done are for the WiFi OOB demo that is only
present on SK. HS is not supported with am64 SK board.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>