conf: branding-processor-sdk: Fix naming of the u-boot.img
All the HS platforms generate the u-boot binary as u-boot.img_HS
but the bootloader expects the binary to be named as u-boot.img

Due to this mismatch, script cannot find the required image to
be added in the WIC image.

Fix this by overwriting the UBOOT_IMAGE variable via branding
conf file.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>