ti-tisdk-makefile: linux-fitimage: Add mew target to create fit images
Add a new target for toplevel makefile that will take the Linux kernel, DTBs
and create a fit image with signed versions of these files.
The FIT image is required to boot a kernel on a secure device.

For creating the FIT image, Use the ITS file that is deployed as part of
prebuilt-images folder.

The fitImage packages the compressed kernel, so add the Image.gz in the
KERNEL_BUILD_CMDS so that the compressed kernel will be built.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>