recipes-graphics: hmi-demo: Update git URL to new repo
Update the git URL to generic repo instead of am62x specific repo.
start_hmi_matrix.sh and start_3d_demo.sh are moved to scripts directory
in the git repo. So remove them and add it as a patch.
On 1920*1200 screen, a white bar is observed at the bottom in fullscreen
mode. Add fix-background-rectangle-size.patch to fix that issue.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>